+++ src/main/java/auth/LoginMemberArgumentResolver.java
@@ -1,5 +1,6 @@
 package auth;
 
+import auth.login.LoginService;
 import lombok.RequiredArgsConstructor;
 import nextstep.support.exception.NoMemberException;
 import org.springframework.core.MethodParameter;
