+++ src/main/java/nextstep/config/AuthConfiguration.java
@@ -2,8 +2,8 @@ package nextstep.config;
 
 import auth.AuthenticateProvider;
 import auth.JwtTokenProvider;
-import auth.LoginController;
-import auth.LoginService;
+import auth.login.LoginController;
+import auth.login.LoginService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
