+++ src/test/java/nextstep/theme/ThemeE2ETest.java
@@ -2,8 +2,8 @@ package nextstep.theme;
 
 import io.restassured.RestAssured;
 import nextstep.AbstractE2ETest;
-import auth.TokenRequest;
-import auth.TokenResponse;
+import auth.login.TokenRequest;
+import auth.login.TokenResponse;
 import nextstep.member.MemberRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
